backend/requirements.txt
openai>=1.0.0
pystray>=0.19.4
pillow>=9.0.0
keyboard>=0.13.5
pyperclip>=1.8.2
requests>=2.28.0
python-dotenv==1.1.0
beautifulsoup4==4.13.4
fabric==3.2.2
requests==2.32.4
Flask==3.1.1
Flask-Bcrypt==1.0.1
flask-cors==6.0.1
flask-sqlalchemy==3.1.1
flask-pymongo==3.0.1
APScheduler==3.11.0
pymongo==4.13.2
SQLAlchemy==2.0.41
psutil==5.9.5
eventlet==0.40.1 
pyyaml>=6.0.1